add homework3
