--- HW1/main.py
@@ -1,2 +1,0 @@
-print('Hello world')
-
