--- HW22/.idea/workspace.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ChangeListManager">
-    <list default="true" id="be52b579-09b2-4b94-bc85-470ce137ba00" name="Changes" comment="">
-      <change beforePath="$PROJECT_DIR$/../.idea/Python.iml" beforeDir="false" afterPath="$PROJECT_DIR$/../.idea/Python.iml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../HW2.1/hw2.1.py" beforeDir="false" afterPath="$PROJECT_DIR$/../HW2.1/hw21.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../HW2.1/test_hw2.1.py" beforeDir="false" afterPath="$PROJECT_DIR$/../HW2.1/test_hw21.py" afterDir="false" />
-    </list>
-    <option name="SHOW_DIALOG" value="false" />
-    <option name="HIGHLIGHT_CONFLICTS" value="true" />
-    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
-    <option name="LAST_RESOLUTION" value="IGNORE" />
-  </component>
-  <component name="Git.Settings">
-    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$/.." />
-  </component>
-  <component name="ProjectId" id="2Nh7yIO4au2AGnQlb1j7clmlrmS" />
-  <component name="ProjectViewState">
-    <option name="hideEmptyMiddlePackages" value="true" />
-    <option name="showLibraryContents" value="true" />
-  </component>
-  <component name="PropertiesComponent"><![CDATA[{
-  "keyToString": {
-    "RunOnceActivity.OpenProjectViewOnStart": "true",
-    "RunOnceActivity.ShowReadmeOnStart": "true",
-    "last_opened_file_path": "C:/Git/irmerdonmz/Phyton"
-  }
-}]]></component>
-  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
-  <component name="TaskManager">
-    <task active="true" id="Default" summary="Default task">
-      <changelist id="be52b579-09b2-4b94-bc85-470ce137ba00" name="Changes" comment="" />
-      <created>1680104712337</created>
-      <option name="number" value="Default" />
-      <option name="presentableId" value="Default" />
-      <updated>1680104712337</updated>
-    </task>
-    <servers />
-  </component>
-</project>
