--- HW22/requirements22.txt
@@ -1,1 +1,0 @@
-pytest
